+++ CHANGELOG.rst
@@ -1,11 +1,14 @@
 Changelog
 ---------
 
-v1.25 (??? 2022)
+v1.25 (Jun 2022)
     * Allow negative dates in `date2date`.
     * Support of microseconds in `date2date`.
     * Assure that `input2array` gives no 0d-array.
     * More `datetime` tests and bug fixes for microseconds and has_year_zero.
+    * Delete unnecessary HDF5 filters in variable definition in `netcdfio`
+      for compatibility with netcdf4 > 1.6.0.
+    * Reduce precision to 14 digits for tests of `logistic2_offset`.
 
 v1.24 (Jun 2022)
     * Removed documentation from `Read the Docs` and moved to Github Pages: